requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==1.4.46
aiofiles==23.2.1
python-multipart==0.0.6
aiohttp==3.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
databases[aiomysql]==0.8.0
pymysql==1.1.0
bcrypt==4.1.1